feat: Split out SimilarityMatrix code
